requirements.txt
cryptography==42.0.5
pycryptodome==3.20.0
cryptography==42.0.5
fastecdsa==2.3.0
questionary==2.1.0
python-dotenv==1.0.1
pycryptodome==3.20.0